Keldon's Newest Update
In this commit I am removing all files not needed and I also
implemented the ability to land on platforms. The ability works but the
animation constraints are a little off also you can still walk through
walls. I will figure out that next. That should be an easy fix