Hero can no longer fall or jump off the screen
So ya I set a function that will kill the hero when the hero gets to
high or too low. This function is supposed to kill the hero but at the
moment it only resets the hero's position to a falling state. I left a
comment, where I would like you to add stuff to make it so he actually
dies.